@@ -1,21 +1,18 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NgModule } from '@angular/core';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
 import { DynamicFormComponent } from './dynamic-form.component';
 import { DynamicFormQuestionComponent } from './dynamic-form-question.component';
 
-@NgModule({
-  imports: [BrowserModule, ReactiveFormsModule, HttpClientModule],
-  declarations: [
-    AppComponent,
-    DynamicFormComponent,
-    DynamicFormQuestionComponent,
-  ],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [
+        AppComponent,
+        DynamicFormComponent,
+        DynamicFormQuestionComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule, ReactiveFormsModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule {}
 
 /*
